requirements.txt
mysql-connector-python
jinja2
Pillow
python-dotenv
pyyaml
requests
playwright
pytest
